+++ apps/heureka/src/components/Services/ServicesList/ServicesList.tsx
@@ -3,26 +3,46 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { useEffect } from "react"
+import React, { useState, useCallback, useEffect } from "react"
 import { DataGrid, DataGridRow, DataGridHeadCell, Pagination, Message } from "@cloudoperators/juno-ui-components"
 import { ServiceListItem } from "./ServiceListItem"
-import { FilterSettings } from "../../common/Filters/types"
-import { useFetchServices } from "../useFetchServices"
 import { EmptyDataGridRow } from "../../common/EmptyDataGridRow/EmptyDataGridRow"
 import { useActions as messageActions } from "@cloudoperators/juno-messages-provider"
+import { ServicePanel } from "../ServicePanel/ServicePanel"
+import { ServiceType } from "../Services"
 
 const COLUMN_SPAN = 6
 
 type ServiceListProps = {
-  filterSettings: FilterSettings
+  loading: boolean
+  error?: string | null
+  services: ServiceType[]
+  currentPage?: number
+  totalNumberOfPages: number
+  goToPage: (page: number | undefined) => void
 }
 
-export const ServicesList = ({ filterSettings }: ServiceListProps) => {
+export const ServicesList = ({
+  loading,
+  error,
+  services,
+  currentPage,
+  totalNumberOfPages,
+  goToPage,
+}: ServiceListProps) => {
   const { addMessage } = messageActions()
+  const [selectedService, setSelectedService] = useState<ServiceType | null>(null)
 
-  const { loading, error, services, currentPage, totalNumberOfPages, goToPage } = useFetchServices({
-    filterSettings,
-  })
+  const handlePanelClose = useCallback(() => {
+    setSelectedService(null)
+  }, [selectedService])
+
+  const handleServiceClick = useCallback(
+    (service: ServiceType) => {
+      setSelectedService(service.name === selectedService?.name ? null : service)
+    },
+    [selectedService]
+  )
 
   useEffect(() => {
     if (error) {
@@ -55,7 +75,9 @@ export const ServicesList = ({ filterSettings }: ServiceListProps) => {
             !loading &&
               !error &&
               services.length > 0 &&
-              services.map((item) => <ServiceListItem key={item.name} item={item} />)
+              services.map((item) => (
+                <ServiceListItem key={item.name} item={item} onClick={() => handleServiceClick(item)} />
+              ))
           }
 
           {
@@ -77,6 +99,7 @@ export const ServicesList = ({ filterSettings }: ServiceListProps) => {
           />
         </div>
       )}
+      {selectedService && <ServicePanel service={selectedService} onClose={handlePanelClose} />}
     </div>
   )
 }